[new-style-fast] Split up Style compiler into Style and GenOptProblem, and organize code in both files
