+++ src/Server.hs
@@ -24,6 +24,7 @@ import System.Random
 import Debug.Trace
 import qualified Shapes  as SD
 import qualified Style   as NS
+import qualified GenOptProblem as G
 import qualified Optimizer as O
 import qualified Data.Map  as M
 import qualified Data.Text                 as T
@@ -37,7 +38,7 @@ import qualified Control.Exception         as Exc (catch, ErrorCall)
 -- COMBAK: remove TagExpr
 
 -- Types used by the server, mainly for translation to JSON
-type ServerState = NS.RState
+type ServerState = G.State
 
 data Feedback = Cmd Command
                 | Drag DragEvent
@@ -80,7 +81,7 @@ wsSendJSONFrame conn frame = WS.sendTextData conn $ encode frame
 --   the Penrose Runtime.
 servePenrose :: String  -- the domain of the server
              -> Int  -- port number of the server
-             -> NS.RState  -- initial state of Penrose Runtime
+             -> G.State  -- initial state of Penrose Runtime
              -> IO ()
 servePenrose domain port initState = do
      putStrLn "Starting Server..."
@@ -122,21 +123,21 @@ application :: ServerState -> WS.ServerApp
 application s pending = do
     conn <- WS.acceptRequest pending
     WS.forkPingThread conn 30 -- To keep the connection alive
-    wsSendJSONList conn (NS.shapesr s)
+    wsSendJSONList conn (G.shapesr s)
     loop conn (O.step s)
 
-loop :: WS.Connection -> NS.RState -> IO ()
+loop :: WS.Connection -> G.State -> IO ()
 loop conn s
-    | NS.optStatus (NS.paramsr s) == NS.EPConverged = do
+    | G.optStatus (G.paramsr s) == G.EPConverged = do
         putStrLn "Optimization completed."
-        putStrLn ("Current weight: " ++ (show $ NS.weight (NS.paramsr s)))
+        putStrLn ("Current weight: " ++ (show $ G.weight (G.paramsr s)))
         wsSendJSONFrame conn (Frame { flag = "final",
-                                shapes = (NS.shapesr s) :: ([SD.Shape Double]) })
+                                shapes = (G.shapesr s) :: ([SD.Shape Double]) })
         processCommand conn s
-    | NS.autostep s = stepAndSend conn s
+    | G.autostep s = stepAndSend conn s
     | otherwise = processCommand conn s
 
-processCommand :: WS.Connection -> NS.RState -> IO ()
+processCommand :: WS.Connection -> G.State -> IO ()
 processCommand conn s = do
     putStrLn "Receiving Commands"
     msg_json <- WS.receiveData conn
@@ -168,54 +169,54 @@ toPolyProperty v = case v of
     SD.FileV x   -> SD.FileV x
     SD.StyleV x  -> SD.StyleV x
 
-updateShapes :: [SD.Shape Double] -> Connection -> NS.RState -> IO ()
+updateShapes :: [SD.Shape Double] -> Connection -> G.State -> IO ()
 updateShapes newShapes conn s =
     let polyShapes = toPolymorphics newShapes
-        uninitVals = map NS.toTagExpr $ NS.shapes2vals polyShapes $ NS.uninitializedPaths s
-        trans' = NS.insertPaths (NS.uninitializedPaths s) uninitVals (NS.transr s)
-        newObjFn = NS.genObjfn trans' (NS.objFns s) (NS.constrFns s) (NS.varyingPaths s)
+        uninitVals = map G.toTagExpr $ G.shapes2vals polyShapes $ G.uninitializedPaths s
+        trans' = G.insertPaths (G.uninitializedPaths s) uninitVals (G.transr s)
+        newObjFn = G.genObjfn trans' (G.objFns s) (G.constrFns s) (G.varyingPaths s)
         news = s {
-            NS.shapesr = polyShapes,
-            NS.varyingState = NS.shapes2floats polyShapes $ NS.varyingPaths s,
-            NS.transr = trans',
-            NS.paramsr = (NS.paramsr s) { NS.weight = NS.initWeight, NS.optStatus = NS.NewIter, NS.overallObjFn = newObjFn }}
-    in if NS.autostep s then stepAndSend conn news else loop conn news
+            G.shapesr = polyShapes,
+            G.varyingState = G.shapes2floats polyShapes $ G.varyingPaths s,
+            G.transr = trans',
+            G.paramsr = (G.paramsr s) { G.weight = G.initWeight, G.optStatus = G.NewIter, G.overallObjFn = newObjFn }}
+    in if G.autostep s then stepAndSend conn news else loop conn news
 
-dragUpdate :: String -> Float -> Float -> WS.Connection -> NS.RState -> IO ()
+dragUpdate :: String -> Float -> Float -> WS.Connection -> G.State -> IO ()
 dragUpdate name xm ym conn s =
     let (xm', ym') = (r2f xm, r2f ym)
         newShapes  = map (\shape ->
             if SD.getName shape == name
                 then SD.setX (SD.FloatV (xm' + SD.getX shape)) $ SD.setY (SD.FloatV (ym' + SD.getY shape)) shape
                 else shape)
-            (NS.shapesr s)
-        news = s { NS.shapesr = newShapes,
-                   NS.varyingState = NS.shapes2floats newShapes $ NS.varyingPaths s,
-                   NS.paramsr = (NS.paramsr s) { NS.weight = NS.initWeight, NS.optStatus = NS.NewIter }}
-   in if NS.autostep s then stepAndSend conn news else loop conn news
+            (G.shapesr s)
+        news = s { G.shapesr = newShapes,
+                   G.varyingState = G.shapes2floats newShapes $ G.varyingPaths s,
+                   G.paramsr = (G.paramsr s) { G.weight = G.initWeight, G.optStatus = G.NewIter }}
+   in if G.autostep s then stepAndSend conn news else loop conn news
 
-executeCommand :: String -> WS.Connection -> NS.RState -> IO ()
+executeCommand :: String -> WS.Connection -> G.State -> IO ()
 executeCommand cmd conn s
     | cmd == "resample" = resampleAndSend conn s
     | cmd == "step"     = stepAndSend conn s
-    | cmd == "autostep" = loop conn (s { NS.autostep = not $ NS.autostep s })
+    | cmd == "autostep" = loop conn (s { G.autostep = not $ G.autostep s })
     | otherwise         = putStrLn ("Can't recognize command " ++ cmd)
 
-resampleAndSend, stepAndSend :: WS.Connection -> NS.RState -> IO ()
+resampleAndSend, stepAndSend :: WS.Connection -> G.State -> IO ()
 resampleAndSend conn s = do
-    let (newShapes, rng') = SD.sampleShapes (NS.rng s) (NS.shapesr s)
-    let uninitVals = map NS.toTagExpr $ NS.shapes2vals newShapes $ NS.uninitializedPaths s
-    let trans' = NS.insertPaths (NS.uninitializedPaths s) uninitVals (NS.transr s)
-                    -- TODO: shapes', rng' = NS.sampleConstrainedState (NS.rng s) (NS.shapesr s) (NS.constrs s)
-    let nexts = s { NS.shapesr = newShapes, NS.rng = rng',
-                    NS.transr = trans',
-                    NS.varyingState = NS.shapes2floats newShapes $ NS.varyingPaths s,
-                    NS.paramsr = (NS.paramsr s) { NS.weight = NS.initWeight, NS.optStatus = NS.NewIter } }
-    wsSendJSONList conn $ fst $ NS.evalTranslation nexts
+    let (newShapes, rng') = SD.sampleShapes (G.rng s) (G.shapesr s)
+    let uninitVals = map G.toTagExpr $ G.shapes2vals newShapes $ G.uninitializedPaths s
+    let trans' = G.insertPaths (G.uninitializedPaths s) uninitVals (G.transr s)
+                    -- TODO: shapes', rng' = G.sampleConstrainedState (G.rng s) (G.shapesr s) (G.constrs s)
+    let nexts = s { G.shapesr = newShapes, G.rng = rng',
+                    G.transr = trans',
+                    G.varyingState = G.shapes2floats newShapes $ G.varyingPaths s,
+                    G.paramsr = (G.paramsr s) { G.weight = G.initWeight, G.optStatus = G.NewIter } }
+    wsSendJSONList conn $ fst $ G.evalTranslation nexts
     loop conn nexts
 
 stepAndSend conn s = do
     let nexts = O.step s
-    wsSendJSONList conn (NS.shapesr nexts :: [SD.Shape Double])
+    wsSendJSONList conn (G.shapesr nexts :: [SD.Shape Double])
     -- loop conn (trRaw "state:" nexts)
     loop conn nexts
